\name{frontier}
\alias{frontier}
\title{Simulated sample from a skew-normal distribution}
\usage{data(frontier)}
\description{
  A sample simulated from the SN(0,1,5) distribution having sample
  index of skewness in the admissible range (-0.9952719,0.9952719) but
  maximum likelihood estimate on the frontier of the paramer space.
}

\format{
 A vector of length 50.
}
\source{
  Generated by a run of \code{rsn(50,0,1,5)}.
}
\examples{
data(frontier)
a <- sn.2logL.profile(y=frontier)
a <- sn.2logL.profile(y=frontier, param.range=c(0.8,1.6,10,30),
        use.cp=FALSE, npts=11)
}
\keyword{datasets}
